conservative linting (#820)
* gofumpt

* gofumpt

* automatic fixes from golangci-lint run ./... --fix

* add source port to events.go

* much more conservative lint

* move the source port to the correct order

* rever changes to democracy.go

* thelper

* allow machine to use all cores

* Revert "allow machine to use all cores"

This reverts commit e536add1a13009cdbe6d6e280c99190c0dabb4d5.

* Revert "Revert "allow machine to use all cores""

This reverts commit 1bb88323e9df0eb222a1965b87043829f1d1f174.

* gofumpt

* use test helpers

* lint after merging main

* lint

---------

+++ tests/difference/core/driver/setup_test.go
@@ -123,8 +123,8 @@ func (s *CoreSuite) TestAssumptionsSetup() {
 	// The offset time is the last committed time, but the SUT is +1 block ahead
 	// because the currentHeader time is ahead of the last committed. Therefore sub
 	// the difference (duration of 1 block).
-	s.Require().Equal(int64(s.offsetTimeUnix), s.time(P).Add(-s.initState.BlockInterval).Unix())
-	s.Require().Equal(int64(s.offsetTimeUnix), s.time(C).Add(-s.initState.BlockInterval).Unix())
+	s.Require().Equal(s.offsetTimeUnix, s.time(P).Add(-s.initState.BlockInterval).Unix())
+	s.Require().Equal(s.offsetTimeUnix, s.time(C).Add(-s.initState.BlockInterval).Unix())
 
 	// The offset height is the last committed height, but the SUT is +1 because
 	// the currentHeader is +1 ahead of the last committed. Therefore sub 1.
